move types into folders (address PR feedback)
