+++ pkg/params/constants.go
@@ -1,4 +1,4 @@
-package types
+package params
 
 // /////
 // Proofs
