+++ pkg/types/fil.go
@@ -2,6 +2,7 @@ package types
 
 import (
 	"fmt"
+	"github.com/filecoin-project/go-shared-types/pkg/params"
 	"math/big"
 	"strings"
 )
@@ -9,7 +10,7 @@ import (
 type FIL BigInt
 
 func (f FIL) String() string {
-	r := new(big.Rat).SetFrac(f.Int, big.NewInt(FilecoinPrecision))
+	r := new(big.Rat).SetFrac(f.Int, big.NewInt(params.FilecoinPrecision))
 	if r.Sign() == 0 {
 		return "0"
 	}
@@ -31,7 +32,7 @@ func ParseFIL(s string) (FIL, error) {
 		return FIL{}, fmt.Errorf("failed to parse %q as a decimal number", s)
 	}
 
-	r = r.Mul(r, big.NewRat(FilecoinPrecision, 1))
+	r = r.Mul(r, big.NewRat(params.FilecoinPrecision, 1))
 	if !r.IsInt() {
 		return FIL{}, fmt.Errorf("invalid FIL value: %q", s)
 	}
